fix folder naming and imports
