+++ src/pages/DailyIncreaseChart/index.js
@@ -1,0 +1,3 @@
+import { drawDailyIncreaseChart } from "./dailyIncrease";
+
+export default drawDailyIncreaseChart;
