+++ src/pages/Header/index.js
@@ -1,0 +1,11 @@
+import { toggleLangPicker } from "./toggleLangPicker";
+import { updateTooltipLang } from "./updateTooltipLang";
+import { drawPageTitleCount } from "./drawPageTitleCount";
+import { drawLastUpdated } from "./drawLastUpdated";
+
+export default {
+  toggleLangPicker,
+  updateTooltipLang,
+  drawPageTitleCount,
+  drawLastUpdated,
+};
