+++ src/pages/OutbreakMap/index.js
@@ -1,0 +1,7 @@
+import { drawMap } from "./drawMap";
+import { drawMapPrefectures } from "./drawMapPrefectures";
+
+export default {
+  drawMap,
+  drawMapPrefectures,
+};
