+++ src/pages/SpreadTrendChart/index.js
@@ -1,0 +1,3 @@
+import { drawTrendChart } from "./spreadTrend";
+
+export default drawTrendChart;
