+++ src/pages/TravelRestrictions/travelRestrictions.js
@@ -1,0 +1,38 @@
+import i18next from "i18next";
+import _ from "lodash";
+
+const travelRestrictionsHelper = (elementId, countries) => {
+  const countryList = [];
+  // Iterate through and render country links
+  _.orderBy(countries, "name", "desc").map((country) => {
+    const name = i18next.t(`countries.${country.name}`);
+    countryList.unshift(
+      `<a href="${country.link}">${country.emoji}${name}</a>`
+    );
+    return true;
+  });
+
+  const banned = document.querySelector(elementId);
+  if (banned) {
+    banned.innerHTML = countryList.join(", ");
+  }
+};
+
+export const drawTravelRestrictions = (ddb) => {
+  travelRestrictionsHelper(
+    "#banned-entry",
+    ddb.travelRestrictions.japan.banned
+  );
+  travelRestrictionsHelper(
+    "#visa-required",
+    ddb.travelRestrictions.japan.visaRequired
+  );
+  travelRestrictionsHelper(
+    "#self-quarantine",
+    ddb.travelRestrictions.japan.selfQuarantine
+  );
+  travelRestrictionsHelper(
+    "#other-restrictions",
+    ddb.travelRestrictions.japan.other
+  );
+};
